requirements.txt
python-nmap
vulners
colorama
socket
